refactor(项目设置): 文件管理的表格数据查询和模块数量统计分成两个接口

+++ backend/services/project-management/src/main/java/io/metersphere/project/dto/ModuleCountDTO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class ModuleCountDTO {
     private String moduleId;
-    private long dataCount;
+    private int dataCount;
 }
